requirements.txt
fastapi
uvicorn[standard]
streamlit
python-dotenv
PyYAML
requests
google-generativeai
pandas
pytest
google-auth-oauthlib
google-api-python-client